+++ src/plugin.cpp
@@ -8,4 +8,6 @@ void init(Plugin* p) {
     p->addModel(modelHaptik);
     p->addModel(modelAxon);
     p->addModel(modelSoma);
+    p->addModel(modelOperon);
+    p->addModel(modelBunnies);
 }